+++ molviewspec/molviewspec/__init__.py
@@ -3,9 +3,13 @@
 __version__ = "1.5.0"
 
 from molviewspec.builder import create_builder
+from molviewspec.mvsx_converter import mvsj_to_mvsx
 from molviewspec.nodes import (
+    MVSJ,
+    MVSX,
     ComponentExpression,
     GlobalMetadata,
+    MVSData,
     PrimitiveComponentExpressions,
     Snapshot,
     SnapshotMetadata,
@@ -13,4 +17,3 @@ from molviewspec.nodes import (
     States,
     validate_state_tree,
 )
-from molviewspec.mvsx_converter import mvsj_to_mvsx